requirements.txt
matplotlib==3.7.1
numpy<2.0.0
plyer==2.1.0
tkcalendar==1.6.1
Pillow==10.0.0